api/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
bcrypt
pandas
google-cloud-storage
docling
redis>=4.5.0
litellm>=1.0.0
async-timeout>=4.0.0
openai
pytest
python-dotenv
requests
PyMuPDF
camelot-py[cv]
pdfplumber
pytesseract
Pillow
reportlab
# Donut - Document Understanding Transformer for tax form extraction
transformers>=4.25.0
torch>=2.0.0
